add patch for printf-style format typechecks (from Robin Barker <rmb1@cise.npl.co.uk>); fixes for problems so identified
p4raw-id: //depot/perl@4836

+++ doio.c
@@ -217,7 +217,7 @@ Perl_do_open9(pTHX_ GV *gv, register char *name, I32 len, int as_raw,
 	if (*type == '|') {
 	    if (num_svs && (tlen != 2 || type[1] != '-')) {
 	      unknown_desr:
-		Perl_croak(aTHX_ "Unknown open() mode '%.*s'", olen, oname);
+		Perl_croak(aTHX_ "Unknown open() mode '%.*s'", (int)olen, oname);
 	    }
 	    /*SUPPRESS 530*/
 	    for (type++, tlen--; isSPACE(*type); type++, tlen--) ;
